backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg[binary]==3.2.4
psycopg-pool==3.2.4
Werkzeug==3.0.1
email-validator==2.1.0.post1
